binary4.03

+++ pr_int.c
@@ -13,7 +13,7 @@ int pr_int(va_list args, t_flag *fl)
 	int x;
 	char s[40] = {0};
 
-	x = va_arg(args, int);
+	x = va_arg(args, long int);
 	num = 0;
 
 	if (x == 0)
